+++ src/dezero/transforms.py
@@ -158,5 +158,5 @@ ToFloat = AsType
 
 
 class ToInt(AsType):
-    def __init__(self, dtype=np.int):
+    def __init__(self, dtype=np.int32):
         self.dtype = dtype
